requirements.txt
numpy
SpeechRecognition
gTTS
opencv-python
openai
keras
tensorflow
Pillow